requirements.txt
Django==4.2.9
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.2.0
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.12.0
python-decouple==3.8
whitenoise==6.6.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
